backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
chromadb==0.4.18
sentence-transformers==2.2.2
pypdf2==3.0.1
openai==1.3.5
python-multipart==0.0.6
google-generativeai==0.3.0
